requirements.txt
watchdog>=4.0.0
Pillow>=10.0.0
pdfplumber>=0.10.0
requests>=2.28.0
imagehash>=4.3.0
